+++ src/components/QueryEditor/VisualQueryEditor/AggregateSection.tsx
@@ -1,3 +1,4 @@
+import { t } from '@grafana/i18n';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { EditorField, EditorFieldGroup, EditorList, EditorRow } from '@grafana/plugin-ui';
 import { QueryEditorExpression, QueryEditorExpressionType, QueryEditorReduceExpression } from 'types/expressions';
@@ -83,7 +84,11 @@ const AggregateSection: React.FC<AggregateSectionProps> = ({
     <div data-testid="aggregate-section">
       <EditorRow>
         <EditorFieldGroup>
-          <EditorField label="Aggregate" optional={true} data-testid={selectors.components.queryEditor.aggregate.field}>
+          <EditorField
+            label={t('components.aggregate-section.label-aggregate', 'Aggregate')}
+            optional={true}
+            data-testid={selectors.components.queryEditor.aggregate.field}
+          >
             <EditorList
               items={aggregates}
               onChange={onChange}