+++ src/components/QueryEditor/VisualQueryEditor/KQLPreview.tsx
@@ -1,3 +1,4 @@
+import { t, Trans } from '@grafana/i18n';
 import React, { useState, useEffect } from 'react';
 import { EditorField, EditorFieldGroup, EditorRow } from '@grafana/plugin-ui';
 import { Button, useStyles2 } from '@grafana/ui';
@@ -23,10 +24,13 @@ const KQLPreview: React.FC<KQLPreviewProps> = ({ query }) => {
   return (
     <EditorRow>
       <EditorFieldGroup>
-        <EditorField label="Query Preview" data-testid={selectors.components.queryEditor.queryPreview.field}>
+        <EditorField
+          label={t('components.kqlpreview.label-query-preview', 'Query Preview')}
+          data-testid={selectors.components.queryEditor.queryPreview.field}
+        >
           <>
             <Button hidden={!hidden} variant="secondary" onClick={() => setHidden(false)} size="sm">
-              show
+              <Trans i18nKey="components.kqlpreview.show">show</Trans>
             </Button>
             <div className={styles.codeBlock} hidden={hidden}>
               <pre className={styles.code}>
@@ -34,7 +38,7 @@ const KQLPreview: React.FC<KQLPreviewProps> = ({ query }) => {
               </pre>
             </div>
             <Button hidden={hidden} variant="secondary" onClick={() => setHidden(true)} size="sm">
-              hide
+              <Trans i18nKey="components.kqlpreview.hide">hide</Trans>
             </Button>
           </>
         </EditorField>